refactor: import the converters from scala.jdk instead of scala.collection
scala.collection.JavaConverters is deprecated in 2.13 in favour of
scala.jdk.CollectionConverters. 31 imports change, in obp-api only, and nothing
else: every conversion at the 59 call sites is the extension-method form
(.asScala / .asJava), and there is not one use of the old named conversions such
as asScalaBuffer, so the new package covers all of them under the same names.

scala.jdk is a 2.13 package, but scala-collection-compat back-ports it to 2.12
and obp-api already declares that dependency - for this very reason, per the
comment on it. obp-commons has no converter imports and so needs nothing.

The list of .asScala call sites matters again at the flip and is worth stating
here: 2.13 makes a bare Seq mean immutable.Seq, while .asScala yields a mutable
Buffer, so these are the places where passing the result to a Seq parameter, or
splatting it with `: _*`, turns into a compile error. Nothing to do about that
now - on 2.12 both spellings behave identically - but that is the population to
look at when the flip produces Seq type errors.

Verification: mvn clean compile test-compile is BUILD SUCCESS, and the full local
suite passes with no FAILED, RUN ABORTED or SUITE ABORTED markers.
