+++ obp-api/src/main/scala/code/api/util/JwsUtil.scala
@@ -18,7 +18,7 @@ import java.time.format.DateTimeFormatter
 import java.time.{Duration, ZoneOffset, ZonedDateTime}
 import java.util
 import scala.collection.immutable.{HashMap, List}
-import scala.collection.JavaConverters._
+import scala.jdk.CollectionConverters._
 
 
 object JwsUtil extends MdcLoggable {
