+++ obp-api/src/main/scala/code/api/v4_0_0/Http4s400.scala
@@ -73,7 +73,7 @@ import java.text.SimpleDateFormat
 import java.util
 import com.networknt.schema.ValidationMessage
 
-import scala.collection.JavaConverters._
+import scala.jdk.CollectionConverters._
 import code.model._   // implicit BankAccountExtended → moderatedBankAccount
 import code.model.dataAccess.AuthUser
 import code.ratelimiting.RateLimitingDI
