+++ obp-api/src/main/scala/code/api/v6_0_0/APIMethods600.scala
@@ -99,7 +99,7 @@ trait APIMethods600
 //import scala.collection.mutable.ArrayBuffer
 //import scala.concurrent.Future
 //import scala.concurrent.duration._
-//import scala.collection.JavaConverters._
+//import scala.jdk.CollectionConverters._
 //import scala.util.Random
 //
 //
