+++ obp-api/src/main/scala/code/api/v7_0_0/Http4s700.scala
@@ -55,7 +55,7 @@ import org.http4s._
 import org.http4s.dsl.io._
 import org.typelevel.ci.CIString
 
-import scala.collection.JavaConverters._
+import scala.jdk.CollectionConverters._
 import scala.collection.mutable.ArrayBuffer
 import scala.concurrent.Future
 import scala.language.{higherKinds, implicitConversions}
