+++ obp-api/src/main/scala/code/obp/grpc/api/ApiProto.scala
@@ -142,7 +142,7 @@ object ApiProto extends _root_.scalapb.GeneratedFileObject {
     _root_.scalapb.descriptors.FileDescriptor.buildFrom(scalaProto, dependencies.map(_.scalaDescriptor))
   }
   lazy val javaDescriptor: com.google.protobuf.Descriptors.FileDescriptor = {
-    import scala.collection.JavaConverters._
+    import scala.jdk.CollectionConverters._
     val javaProto = com.google.protobuf.DescriptorProtos.FileDescriptorProto.parseFrom(ProtoBytes)
     // Filter ObpService to expose only getBanks. The other methods
     // (getPrivateAccountsAtOneBank, getBankAccountsBalances,
