+++ obp-api/src/test/scala/code/api/v7_0_0/V7ResourceDocsAggregationTest.scala
@@ -8,7 +8,7 @@ import org.json4s.JsonAST.{JArray, JObject, JString}
 import com.openbankproject.commons.util.JsonAliases.parse
 import org.scalatest.Tag
 
-import scala.collection.JavaConverters._
+import scala.jdk.CollectionConverters._
 import com.openbankproject.commons.util.JsonAliases.RichJField
 
 /**
